+++ package.json
@@ -14,12 +14,13 @@
     "bootstrap": "5.1.3",
     "chart.js": "^3.6.0",
     "chartjs-adapter-moment": "^1.0.0",
+    "chartjs-plugin-zoom": "^1.1.1",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
-    "chartjs-plugin-zoom": "^1.1.1",
     "moment": "^2.29.1",
     "morgan": "^1.10.0",
+    "mysql2": "^2.3.3-rc.0",
     "nodemon": "^2.0.15",
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.2",
@@ -32,6 +33,8 @@
     "react-scripts": "4.0.3",
     "react-scroll": "^1.8.4",
     "sass": "^1.43.4",
+    "sequelize": "^6.9.0",
+    "sequelize-cli": "^6.3.0",
     "styled-components": "^5.3.3",
     "switch": "^0.0.0",
     "web-vitals": "^1.1.2"
@@ -39,6 +42,7 @@
   "scripts": {
     "start": "react-scripts start",
     "start:server": "nodemon ./server/server.js",
+    "start:DBserver": "nodemon ./DB_server/server.js",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
