+++ src/components/ChartSection/Coingecko/CoinList.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Col, FormControl, InputGroup } from "react-bootstrap";
 import { coingecko, coingeckoUrl } from "../../../apis/configs";
-import useAxios from "../../../apis/useAxios";
+import useAxios from "../../../hooks/useAxios";
 import Coin from "./Coin";
 
 const CoinList = () => {
